% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_casting.R
\name{display_palette}
\alias{display_palette}
\title{Display a palette using fillable shapes}
\usage{
display_palette(
  fill,
  n,
  pal_name,
  colour = "grey30",
  color = colour,
  shape = "jar",
  shape_size = 1,
  label_size = shape_size * 3.5
)
}
\arguments{
\item{fill}{The colour of the shape fill.}

\item{n}{Integer number of colours in the palette.}

\item{pal_name}{A character string for the name of the palette.}

\item{colour}{The colour of the shape outline.}

\item{color}{The color of the shape outline.}

\item{shape}{A character string for the name of the shape, e.g. "jar".}

\item{shape_size}{Modifies the appearance of the shape.}

\item{label_size}{Modifies the appearance of the shape's label.}
}
\value{
A ggplot2 object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Creates a visualisation of a chosen palette with each colour in the
selected fillable shape.
}
\examples{
display_palette(c("red", "blue"), n = 2, pal_name = "Example")
}
