% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfoundry-package.R
\docType{package}
\name{ggfoundry-package}
\alias{ggfoundry}
\alias{ggfoundry-package}
\title{ggfoundry: Shape Foundry & Geom for 'ggplot2'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A 'ggplot2' extension that supports arbitrary hand-crafted colourable & fillable shapes. New shapes may be feature requested via a Github issue.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cgoo4/ggfoundry}
  \item \url{https://cgoo4.github.io/ggfoundry/}
  \item Report bugs at \url{https://github.com/cgoo4/ggfoundry/issues}
}

}
\author{
\strong{Maintainer}: Carl Goodwin \email{carl.goodwin@quantumjitter.com} [copyright holder]

Authors:
\itemize{
  \item Carl Goodwin \email{carl.goodwin@quantumjitter.com} [copyright holder]
}

}
\keyword{internal}
