% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_casting.R
\name{shapes_cast}
\alias{shapes_cast}
\title{Get the names of available shapes}
\usage{
shapes_cast()
}
\value{
A data frame of available sets and shapes.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Create a data frame of available shapes and associated sets. This may be
filtered and used as a vector of strings in \code{scale_shape_manual()}.
}
\examples{
# Returns a data frame of available shapes
shapes_cast()
}
